Starting with gdistance:

```
# this PR
julia> @benchmark LightGraphs.gdist2($g,1)
BenchmarkTools.Trial:
  memory estimate:  12.13 MiB
  allocs estimate:  11
  --------------
  minimum time:     182.685 ms (0.00% GC)
  median time:      189.283 ms (0.00% GC)
  mean time:        191.267 ms (0.00% GC)
  maximum time:     216.101 ms (0.00% GC)
  --------------
  samples:          26
  evals/sample:     1

# current master
julia> @benchmark LightGraphs.gdistances($g,1)
BenchmarkTools.Trial:
  memory estimate:  9.00 MiB
  allocs estimate:  6
  --------------
  minimum time:     228.320 ms (0.00% GC)
  median time:      238.889 ms (0.00% GC)
  mean time:        242.413 ms (0.57% GC)
  maximum time:     272.067 ms (0.00% GC)
  --------------
  samples:          21
  evals/sample:     1

# this PR
julia> @benchmark LightGraphs.gdist2($sg,1)
BenchmarkTools.Trial:
  memory estimate:  12.13 MiB
  allocs estimate:  11
  --------------
  minimum time:     126.385 ms (0.00% GC)
  median time:      135.065 ms (0.00% GC)
  mean time:        143.928 ms (0.00% GC)
  maximum time:     295.152 ms (0.00% GC)
  --------------
  samples:          35
  evals/sample:     1

# current master
julia> @benchmark LightGraphs.gdistances($sg,1)
BenchmarkTools.Trial:
  memory estimate:  9.00 MiB
  allocs estimate:  6
  --------------
  minimum time:     146.291 ms (0.00% GC)
  median time:      171.028 ms (0.00% GC)
  mean time:        177.256 ms (0.93% GC)
  maximum time:     302.972 ms (0.00% GC)
  --------------
  samples:          29
  evals/sample:     1
```