@sbromberger 2017-11-10
cc @jpfairbanks, @somil55, and/or @juliohm for a review.