@sbromberger 2017-11-14
I will merge this tomorrow unless there are any objections.